+++ clang-tools-extra/clang-tidy/modernize/CStringToStdFilenameCheck.h
@@ -1,0 +1,36 @@
+//===--- CStringToStdFilenameCheck.h - clang-tidy ---------------*- C++ -*-===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_MODERNIZE_CSTRINGTOSTDFILENAMECHECK_H
+#define LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_MODERNIZE_CSTRINGTOSTDFILENAMECHECK_H
+
+#include "../utils/TransformerClangTidyCheck.h"
+
+namespace clang::tidy::modernize {
+
+/// FIXME: Write a short description.
+///
+/// For the user-facing documentation see:
+/// http://clang.llvm.org/extra/clang-tidy/checks/modernize/C-string-to-std-filename.html
+class CStringToStdFilenameCheck : public ::clang::tidy::utils::TransformerClangTidyCheck {
+public:
+        CStringToStdFilenameCheck(StringRef Name, ClangTidyContext *Context);
+
+        bool isLanguageVersionSupported(const LangOptions &LangOpts) const override {
+            return LangOpts.CPlusPlus17;
+        }
+
+        /*
+  void registerMatchers(ast_matchers::MatchFinder *Finder) override;
+  void check(const ast_matchers::MatchFinder::MatchResult &Result) override;
+        */
+};
+
+} // namespace clang::tidy::modernize
+
+#endif // LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_MODERNIZE_CSTRINGTOSTDFILENAMECHECK_H